requirements.txt
Django==3.2.12
python-dotenv
stripe
